requirements.txt
Django==5.0.4
psutil==5.9.1
requests==2.28.2
xlrd==1.2.0
openpyxl==3.1.2
netifaces==0.11.0
pillow==9.5.0
opencv-python==4.5.4.60
numpy==1.26.4